start&controllerボタン追加
