+++ server/api/game/_gameId@string/validators.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod'
+import { defineValidators } from './$relay'
+
+export default defineValidators(() => ({
+  params: z.object({ gameId: z.string() })
+}))
