+++ src/pages/welcomepage/index.module.css
@@ -1,0 +1,22 @@
+.title {
+  font-size: 80px;
+  text-align: center;
+  font-weight: bold;
+}
+
+.tasks {
+  width: 300px;
+  padding: 0;
+  margin: 20px auto 0;
+  list-style-type: none;
+  text-align: left;
+}
+
+.tasks > li {
+  margin-top: 10px;
+  border-bottom: 1px solid #eee;
+}
+
+.deleteBtn {
+  float: right;
+}
